// client/src/store/action-types/productActionType.ts
export enum ProductActionType {
  GET_ALL_PRODUCTS = 'get_all_products',
  GET_PRODUCT_BY_ID = 'get_product_by_id',
  CREATE_PRODUCT = 'create_product',
  UPDATE_PRODUCT = 'update_product',
  DELETE_PRODUCT = 'delete_product',
  PRODUCT_SUCCESS = 'product_success',
  PRODUCT_ERROR = 'product_error',
}